Complete firebase-react-auth video
Found an issue where a weird behavior would occur when
attempting to update both email and password at the same time.
Only the password would update and it would navigate to a new path.
I'm not sure what causes it, but this issue is worth revisiting
to further understand the logic.